Cast numbers to numeric type instead of integer

+++ packages/data-driver-postgres/src/utils/conditions/index.test.ts
@@ -238,6 +238,6 @@ test('Convert logical filter on json value like for o2a', () => {
 			parameterIndex2 + 1
 		} AND CAST("t${tableIndex}"."${jsonColumnName}" -> $${parameterIndex3 + 1} -> $${
 			parameterIndex4 + 1
-		} AS INTEGER) = $${parameterIndex5 + 1}`,
+		} AS numeric) = $${parameterIndex5 + 1}`,
 	);
 });
